--bug=876634951 兼容前端使用的扫描方案类型

+++ src/backend/scanner-executor/biz-scanner-executor/src/main/kotlin/com/tencent/bkrepo/scanner/executor/configuration/ScannerExecutorProperties.kt
@@ -55,5 +55,5 @@ data class ScannerExecutorProperties(
     /**
      * 扫描器日志最大行数
      */
-    var maxScannerLogLines: Long = 200L
+    var maxScannerLogLines: Int = 200
 )
